change policy behavior to not decide to rescale if missing key metric drivers

+++ springline/tests/phases/fixtures.rs
@@ -11,7 +11,7 @@ lazy_static! {
     pub static ref DT_1_TS: i64 = DT_1.timestamp();
 }
 
-pub fn make_test_item(_timestamp: &DateTime<Utc>, records_in_per_sec: f64, inbox_lag: f64) -> Telemetry {
+pub fn make_test_item(_timestamp: &DateTime<Utc>, records_in_per_sec: f64, inbox_lag: i64) -> Telemetry {
     let item = maplit::hashmap! {
         // "timestamp".to_string() => Timestamp::from_datetime(&timestamp).into(),
         MC_FLOW__RECORDS_IN_PER_SEC.to_string() => records_in_per_sec.into(),
